product exceptions

+++ src/main/java/org/iths/ecommercebe/Service/ProductService.java
@@ -1,5 +1,6 @@
 package org.iths.ecommercebe.Service;
 
+import org.iths.ecommercebe.Exceptions.ProductNotFoundException;
 import org.iths.ecommercebe.Model.Product;
 import org.iths.ecommercebe.Repository.ProductRepository;
 import org.springframework.stereotype.Service;
@@ -19,14 +20,25 @@ public class ProductService {
     }
 
     public Product getProductById(Long id) {
-        return productRepository.findById(id).orElse(null);
+        return productRepository.findById(id).orElseThrow(() -> new ProductNotFoundException("Product (id=" + id + ") not found"));
     }
 
     public Product createProduct(Product product) {
         return productRepository.save(product);
     }
 
+    public Product updateProduct(Long id, Product updatedProduct) {
+        if (!productRepository.existsById(id)) {
+            throw new ProductNotFoundException("Product (id=" + id + ") not found");
+        }
+        updatedProduct.setId(id);
+        return productRepository.save(updatedProduct);
+    }
+
     public void deleteProduct(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ProductNotFoundException("Product (id=" + id + ") not found");
+        }
         productRepository.deleteById(id);
     }
 }
